Added lab 9 files. Updated Readme.md
Lab 9 on Lviv Polytechnic IoT course.

+++ src/main/java/ua/lviv/iot/stationery/models/ArtItem.java
@@ -1,6 +1,7 @@
 package ua.lviv.iot.stationery.models;
 
 import lombok.Data;
+import lombok.ToString;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
@@ -16,11 +17,13 @@ public class ArtItem extends Item {
     private String color;
     private DrawingTools drawToolType;
 
-    public ArtItem(final String name, final float weight, final float price,
-                   final float rating, final CustomerType customerType,
+    public ArtItem(final String name, final double weight, final double price,
+                   final double rating, final CustomerType customerType,
                    final String thisColor, final DrawingTools thisDrawToolType) {
         super(name, weight, price, rating, customerType);
         this.color = thisColor;
         this.drawToolType = thisDrawToolType;
     }
 }
+
+
